Control.Monad.Freer.Extras.Log

Documentation

This module provides effects and handlers for structured logging and tracing.

data LogMsg a r where

Constructors

LMessage :: LogMessage a -> LogMsg a ()

data LogLevel

The severity level of a log message See https://en.wikipedia.org/wiki/Syslog#Severity_level

Constructors

Debug

Info

Notice

Warning

Error

Critical

Alert

Emergency

data LogMessage a

Constructors

LogMessage

Fields

_logLevel :: LogLevel

_logMessageContent :: a

logLevel :: forall a. Lens' (LogMessage a) LogLevel

logMessageContent :: forall a a. Lens (LogMessage a) (LogMessage a) a a

logMessage :: LogLevel -> Prism' (LogMessage a) a

logDebug :: forall a effs. Member (LogMsg a) effs => a -> Eff effs ()

logInfo :: forall a effs. Member (LogMsg a) effs => a -> Eff effs ()

logWarn :: forall a effs. Member (LogMsg a) effs => a -> Eff effs ()

logError :: forall a effs. Member (LogMsg a) effs => a -> Eff effs ()

Modifying logs

mapLog :: forall a b effs. Member (LogMsg b) effs => (a -> b) -> LogMsg a ~> Eff effs

Re-interpret a logging effect by mapping the log messages. (Does the same thing as contramap for Trace)

mapMLog :: forall a b effs. Member (LogMsg b) effs => (a -> Eff effs b) -> LogMsg a ~> Eff effs

Re-interpret a logging effect by mapping the log messages. Can use other effects.

Running logs

handleWriterLog :: forall a f effs. (Member (LogMsg a) effs, Traversable f) => (a -> LogLevel) -> Eff (Writer (f a) ': effs) ~> Eff effs

Re-interpret a Writer effect by writing the events to the log

handleLogIgnore :: Eff (LogMsg a ': effs) ~> Eff effs

Ignore all log messages.

handleLogTrace :: Pretty a => Eff (LogMsg a ': effs) ~> Eff effs

Write the log to stdout using trace

handleLogWriter :: forall a w effs. Member (Writer w) effs => AReview w (LogMessage a) -> LogMsg a ~> Eff effs

Re-interpret a Log effect with a Writer

renderLogMessages :: forall a effs. (Member (LogMsg Text) effs, Pretty a) => LogMsg a ~> Eff effs

Pretty-print the log messages

Observing

data LogObserve a r where

Constructors

ObserveBefore :: a -> LogObserve a ObservationHandle

ObserveAfter :: Maybe a -> ObservationHandle -> LogObserve a ()

data ObservationHandle

An abstract type used to tie the beginning and end of observations together.

data Observation v s

An observation with measurements before and after running an action.

Constructors

Observation

Fields

obsLabelStart :: v

Call-site information about the start of the observation

obsStart :: s

Measurement taken before running the action

obsLabelEnd :: Maybe v

Call-site information about the end of the observation

obsExit :: ExitMode

ExitMode of the action.

observeBefore :: forall a. forall effs. Member (LogObserve a) effs => a -> Eff effs ObservationHandle

observeAfter :: forall a. forall effs. Member (LogObserve a) effs => Maybe a -> ObservationHandle -> Eff effs ()

Combinators

surround :: forall v a effs. Member (LogObserve v) effs => v -> Eff effs a -> Eff effs a

Write a log message before and after an action. Consider using observeBefore and observeAfter directly if you need more control over the values that are observed at the call site.

surroundDebug :: Member (LogObserve (LogMessage v)) effs => v -> Eff effs a -> Eff effs a

surroundDebug = surround Debug

surroundInfo :: Member (LogObserve (LogMessage v)) effs => v -> Eff effs a -> Eff effs a

surroundInfo = surround Info

surroundWarn :: Member (LogObserve (LogMessage v)) effs => v -> Eff effs a -> Eff effs a

surroundWarn = surround Warn

Handlers

handleObserveLog :: forall effs. Member (LogMsg Text) effs => Eff (LogObserve (LogMessage Text) ': effs) ~> Eff effs

Interpret the LogObserve effect by logging a "start" message before the action and an "end" message after the action.

handleObserve

Arguments

:: forall v s effs. (v -> Eff effs s)
How to get the current s

-> (Observation v s -> Eff effs ())

-> Eff (LogObserve v ': effs) ~> Eff effs

Handle the LogObserve effect by recording observations s before and after the observed action, and turning them into 'LogMessage (Observation s)' values.